The compare editor does not show the proper content
This also fixes the issue where the comment flag is not properly
displayed (this pb resulted from the addition of the lazy loading).
In order to fix this, we are adding two new attributes to the model. One
called isComment to the ChangeMessage class, and one called isCommented
to the RevisionInfo. Both values are derived from the text in the
message.

Change-Id: I79c41d3167566dda41a02b3045802980b50ce692