Bug 490009 - [Compare] Some files are not presented
Rather than diffing the list of files contained in each revision and
having to deal with the case where some files disappeared between two
revisions. We simply use the appropriate call to get the diff between
two revisions.

Because of this, we no longer need the button to filter the modified
files between two revisions

Change-Id: I4ea6a3acaafa2ce17b763ffc22955e7a8fc7f913