Add SourceReferences to plugins
This allows the user to clone and checkout the source code of EGerrit
without knowing where the repository is. See:
https://wiki.eclipse.org/PDE/UI/SourceReferences

One use case is to right-click on a plug-in under the Plug-in
dependencies in Package Explorer view and select Import from Repository.
Note that for this EGit needs to be installed.

Another use case is to go to
File > Import > Plug-In Development > Plug-ins and Fragments then select
Projects from a repository.

What this patch does is to let Tycho generate the source references in
the MANIFEST.MF of the built jars. So to test this patch, one can build
the update site locally and inspect the manifests of the (non-source)
jars.

Also see:
http://download.eclipse.org/releases/staging/buildInfo/reporeports/reports/esdata.txt

Change-Id: I50aa1ff1e18bf77c62f80a7d823ee4b060a9198b
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>