Don't block the UI thread when fetching proposed reviewers
The first request to the Gerrit server to get the list of proposed
reviewers takes quite a long time.  The current code makes the call to
fetch reviewers on the UI thread and the user cannot do anything until
the call completes, not even type extra letters to make the reviewer
search more specific.

This patch moves the call to the server to a separate job thus allowing
the UI thread to not be blocked.  With this change, multiple queries to
the server can happen, but only the last one will be used.
 
Change-Id: I4ae367d42ab5b7cb304b20be267fb665e4389efd
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>