Allow test to run with Jenkins
When running the EGerritCorePluginTest, it re-initialize the bundle context
and the following unit test, when requesting the plugin context, it received
a null context pointer. I just make sure to maintain available the plugin
context after the EGerritCorePluginTest has terminated to fix the issue.
We were able to execute the unit test individually in eclipse,
but were failing when running in a batch from maven


Change-Id: Ie459bf6c863c52c506d8caf564ed4904e399b6e0
Signed-off-by: Jacques Bouthillier <lmcbout@gmail.com>