Provide order for search completion to increase usefulness
When using the search completion of the dashboard I found that the
completions I was most interested in were often low on the list of
completion so I had to press the down key many times to select it.

This patch provides what I believe is a better order for the completion
suggestions.

Note that webGerrit shows the choices in alphatbetical order.  I'm not
sure if being (hopefully) better but different is a problem.

Change-Id: I353e9f4cb09141a87ab0db7d855137ffdc36a4ed
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>