Other revisions are not loaded
The trick we used to decide whether a review was fully loaded got
defeated. This patch introduces a new flag called "loadingLevel" which
allows to track more adequately whether a review has been fully loaded
or not. The flag is set to 0 by default and gets moved to 1 when all the
information about a review are loaded.

Change-Id: Ib96d99fbfe6b340983e7309354d929d6173f4240