Bug 506087 - Null pointer when opening an editor
The problem was not showing all the time because it is timing related.
If the messages have reloaded before the UI is shown, then no problem
occurred. However if the UI was in the process of showing and then the
messages got refreshed, the NPE could occur. 

The fix does two things:
- Handle the exception since it can happen
- Don't unnecessarily refresh the messages if those have not changed.

Change-Id: Ifed5f86c749db70affce40de8a234e448ce3fc8c
Signed-off-by: Jacques Bouthillier <lmcbout@gmail.com>
Signed-off-by: Pascal Rapicault <pascal@rapicorp.com>