Tuleap #675 - Executing query from the dashboard blocks the UI thread
The query was already being done in the background but the retrieval of
the server version was done on the UI thread which caused the problem. 

The version check is now done when is now done everytime we are doing a
query

The code also cleans up the handling of unauthorized queries.

Change-Id: I7effb2ddf8cab1dfa3b142a85d3bbed17fcf50f1
Signed-off-by: Pascal Rapicault <pascal@rapicorp.com>