Tuleap #689 - Compare editor does not show all differences properly
The logic handling the creation of the input for the comparison of a two
revisions has been fixed and completely rewritten to explicitly handle
all the cases that could occur rather than trying to be smart and share
code. This improves the code readability.


The presentation of the status has also been fixed because it was not
handling all the statuses.


Change-Id: I0a2ffd99942527a0ae36dcd0029370b43968c9d9