Bring dashboard view to front on updates
The query of the dashboard view can be changed from outside, e.g. when
clicking onto one of the links on the message page of the
ChangeDetailEditor. Therefore the view needs to be made active, since it
might not be active yet.

Otherwise there is no visible reaction when one of the links is clicked
by the user in the ChangeDetailEditor.

Change-Id: I085871b3e6a96add8ce9bd580910f17609f1a50f
Signed-off-by: Michael Keppler <michael.keppler@gmx.de>