Make the tests more resilient to slow gerrit servers
The tests assumed the test server was running locally, and expected the server response to have been received 
when the test was checking the assertions. However when running against a remote server (e.g. when we are running the tests
from the foundation server, the test server is hosted on a google machine), this is no longer true, and we need to introduce delays
before checking the assertions.

Change-Id: I5d405956db625587f314683d1258e9aa52bb113f