Force reload does not cause an update to the last updated date
The detection of a change in the review is done by watching the last
updated date.
When we are forcing a reload, this date needs to be forced-modified in
order for the listeners watching this date to be notified.

Also, in order to avoid a number of spurious reloads, the listener
reloading the basic information is disabled when doing a force reload.


Change-Id: I1f2fb6a696ab5650f25023ab15e9b1dede4f9f72
Signed-off-by: Pascal Rapicault <pascal@rapicorp.com>