NPE using reply from the compare editor
This bug resulted from the handling of the swap button in Oxygen.
GMI.getRightSide() returned what is given as input to the compare
editor, however this does not represent the value shown on the right in
the title bar of the compare editor.
We needed to introduce a new method returning what is actually shown so
the users of the input don't have to worry about whether the editor is
shown in mirror mode or not.

Change-Id: Icd4d732c6c8fa5977d4d9abcacb60b0f7e326a1b
Signed-off-by: Pascal Rapicault <pascal@rapicorp.com>