Clarify navigation in compare editor
Before this patch, the presentation used for the selection of patchsets
(ps3 / ps2) lead the user to understand that the editors would show
patchsets in the same order: ps3 on the left and ps2 on the right.
This was right, but with the introduction of the swap button and our
attempt at reacting to the swap action by reordering the title in the
upper section of the editor caused additional troubles and unclear
behaviours. For example, what is the patchset that the reply button acts
on, which patchset does the "next comment" button acts on, etc.
In order to clarify the situation,  we have done the following changes:

- In the upper section, change the presentation of patchsets to explicit
the direction of the comparison (since ps3 vs ps1 is different than ps1
vs ps3).
- No longer swap the patchset in the upper section when the user hits
the swap button. The swap button only affects the editor area of the
compare section.
- The next / previous comment button applies to the editor showing the
patchset shown first in the title.

Change-Id: I9474bb200b487363335cab853e4386baad7362b2
Signed-off-by: Pascal Rapicault <pascal@rapicorp.com>