Make sure that compare editor does syntax coloring
The problem was rooted in the inability for the compare editor (and
content type manager) to detect the appropriate content type. This was
because the name of element (ITypedElement#getName) included the name of
the patchset and additional information (like commit id) which prevented
the content type to work.
The patch consits in making sure that getName() only returns the file
name, adding a new method to get the user readable name to be presented
in the title of the sub-editors.

Change-Id: Icb0239f52876bf6a55ad07bc94018956d345703b